refined DELETE and POST method logics of bookings

+++ app/api/bookings/[id]/route.js
@@ -1,15 +1,23 @@
+import { auth } from "@/auth";
 import clientPromise from "@/lib/mongodb";
 import { ObjectId } from "mongodb";
 
 export async function DELETE(req, context){
+    const session = await auth();
+    if(!session || !session.user) return Response.json({error: "Unauthorized"}, {status: 401});
     const {id} = await context.params;
 
     const client = await clientPromise;
     const db = client.db("carDoctor");
 
     let query;
-    if(ObjectId.isValid(id)) query = {_id: new ObjectId(id)};
-    else query = {booking_id: id};
+    if(session.user.role === "admin"){
+        query = {_id: new ObjectId(id)};
+    }else{
+        query = {_id: new ObjectId(id), 
+            "customer.email": session.user.email
+        }
+    }
 
     const result = await db.collection("bookings").deleteOne(query);
